[feat] a lot of things

+++ preprocessing/msst_denoise.py
@@ -1,9 +1,10 @@
 from pathlib import Path
 import subprocess
 import shutil
+import os
 
 def _read_config(msst_config:dict):
-    msst_root = Path(Path.cwd() / msst_config["msstRoot"])
+    msst_root = Path(__file__).resolve().parent / msst_config["msstRoot"]
     return {
         "python":       msst_root / msst_config["workenv"] / "python.exe",
         ".py":          msst_root / msst_config["infer_py"],
@@ -25,7 +26,8 @@ def msst_denoise(input: Path, output: Path, msst_config: dict):
         input_folder = tmp_folder
 
     print("------------------MSST-Denoise------------------")
-    subprocess.run([
+    # print(configs)
+    res = subprocess.run([
         configs["python"], "-u", configs[".py"],
         "--config_path",        str(configs["config"]),
         "--start_check_point",  str(configs["model_ckpt"]),
@@ -33,14 +35,15 @@ def msst_denoise(input: Path, output: Path, msst_config: dict):
         "--store_dir",          str(output),
         "--model_type",         str(configs["model_type"]),
         "--device_ids",         "0",
-    ])
+    ], stderr=subprocess.PIPE)
+
+    if res.returncode != 0:
+        print(res.stderr.decode("utf-8"))
+        raise RuntimeError("MSST-Denoise failed")
 
     for audio in list(output.iterdir()):
-        if not audio.stem.endswith("_vocals"): continue
-        new_name = output.joinpath(audio.stem.rstrip("_vocals") + ".wav")
-        if audio == new_name: continue
-        if new_name.exists(): new_name.unlink()
-        audio.rename(new_name)
+        if not audio.is_dir(): continue
+        shutil.move(audio / "vocals.wav", output / (audio.stem + ".wav"))
 
     if input.is_file():
         shutil.rmtree(input_folder)
@@ -51,6 +54,6 @@ def msst_denoise(input: Path, output: Path, msst_config: dict):
 if __name__ == "__main__":
     import json
     config = json.load(open("./config.json"))["denoiseConfig"]
-    Path("./tmp/denoised/test_vocals.wav").rename(Path("./tmp/denoised/test.wav"))
-    # denoise(Path("./test"), Path("./denoised_result"), config)
-    # denoise(Path("./test/test.wav"), Path("./test/denoised_result"), config)
+    # Path("./tmp/denoised/test_vocals.wav").rename(Path("./tmp/denoised/test.wav"))
+    msst_denoise(Path("./test"), Path("./denoised_result"), config)
+    # msst_denoise(Path("./test/test.wav"), Path("./test/denoised_result"), config)
